backend/requirements.txt
aiocache==0.12.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.10.0
-e git+https://github.com/DevikaReddi/skill-sync-pro.git@a12413fb2340378f72df9787dfc53b25d8de486a#egg=app&subdirectory=backend
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
black==23.12.1
blis==0.7.11
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
coverage==7.10.5
cryptography==45.0.6
cymem==2.0.11
defusedxml==0.7.1
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.109.0
filelock==3.19.1
flake8==7.0.0
fonttools==4.59.2
fpdf2==2.7.7
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gevent==25.5.1
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
limits==5.5.0
Mako==1.3.10
marisa-trie==1.3.1
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
mypy==1.8.0
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
preshed==3.0.10
prometheus-client==0.19.0
propcache==0.3.2
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
python-engineio==4.12.2
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.11.0
PyYAML==6.0.2
redis==5.0.1
regex==2025.7.34
requests==2.32.5
requests-file==2.1.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.2.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.5.1
starlette==0.35.1
sympy==1.14.0
thinc==8.2.5
threadpoolctl==3.6.0
tldextract==5.3.0
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.4
triton==3.4.0
typer==0.9.4
types-python-dateutil==2.8.19.14
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.27.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.3.4
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
zope.event==5.1.1
zope.interface==7.2
